requirements.txt
flask
python-dotenv
xai_sdk
websockets
gunicorn
